Change project's structure and adding more
Project's structure is now changed. For each coordinate system there is
a converter class of its own.
Adding file converter feature. Allow user to convert coordinates through
a text file where all the data is resided. The location to the file
should be given as a source parameter in GET method, also with the
encode of the file.